♻️ Split index and main
