+++ test/fail-on-warning.test.ts
@@ -1,7 +1,8 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
-import { setupOctokitMocks, runAction } from "./utils";
+import { setupOctokitMocks } from "./utils";
 import type * as CoreType from "@actions/core";
 import type { Octokit } from "@octokit/rest";
+import { main } from "../src/main";
 
 vi.mock("@actions/core");
 vi.mock("@actions/github", () => ({
@@ -60,7 +61,7 @@ describe("fail-on-warning behavior", () => {
       return "";
     });
 
-    await runAction();
+    await main();
 
     if (c.shouldFail) {
       expect(core.setFailed).toHaveBeenCalledWith(
